Code-Cleaning, smaller Fixes, project reorganization
