+++ src/app/app.module.ts
@@ -1,16 +1,20 @@
 import { BrowserModule } from '@angular/platform-browser';
 import { ErrorHandler, NgModule } from '@angular/core';
+import { HttpClientModule } from '@angular/common/http';
+
 import { IonicApp, IonicErrorHandler, IonicModule } from 'ionic-angular';
 import { SplashScreen } from '@ionic-native/splash-screen';
 import { StatusBar } from '@ionic-native/status-bar';
+import { LocalNotifications } from '@ionic-native/local-notifications';
+import { NativeGeocoder } from '@ionic-native/native-geocoder';
+import { IonicStorageModule } from '@ionic/storage';
+
+import { LeafletModule } from "@asymmetrik/ngx-leaflet";
 
 import { openSenseApp } from './app.component';
 import { ApiProvider } from '../providers/api/api';
-import { HttpClientModule } from '@angular/common/http';
-import { LocalNotifications } from '@ionic-native/local-notifications';
-import { NativeGeocoder } from '@ionic-native/native-geocoder';
-import { LeafletModule } from "@asymmetrik/ngx-leaflet";
-import { IonicStorageModule } from '@ionic/storage';
+import { helpers } from '../pages/sensify/js/helpers';
+import { validation } from '../pages/sensify/js/validation';
 
 @NgModule({
     declarations: [
@@ -22,7 +26,6 @@ import { IonicStorageModule } from '@ionic/storage';
         IonicModule.forRoot(openSenseApp),
         LeafletModule,
         IonicStorageModule.forRoot()
-
     ],
     bootstrap: [IonicApp],
     entryComponents: [
@@ -35,6 +38,8 @@ import { IonicStorageModule } from '@ionic/storage';
         ApiProvider,
         LocalNotifications,
         NativeGeocoder,
+        helpers,
+        validation
     ]
 })
 export class AppModule { }
